Fix bugs about scheduling and virtual file system

+++ arch/x86/types.h
@@ -65,7 +65,7 @@ typedef unsigned long long int	UINT64;
 
 typedef uint64_t 		size_t;
 typedef int64_t			ssize_t;
-typedef uint64_t		id_t;
+typedef int32_t			id_t;
 typedef uint8_t			atomic_t;
 typedef int32_t			nice_t;
 typedef uint16_t		wchar;
